+++ controllers/shareController.js
@@ -2,15 +2,47 @@ const asyncHandler = require("express-async-handler");
 const { matchedData } = require("express-validator");
 const { validateShareDuration } = require("../validators/validators");
 const {
+  prisma,
   traverseNestedFolders,
   updateFolderExpiresAt,
 } = require("../db/prisma");
+const { isExpired } = require("../utils/utils");
 
 const shareController = {
   getSharedRoute: asyncHandler(async (req, res) => {
     console.log("getShare running...");
     console.log("req.params:", req.params);
+    const { folderID } = req.params;
+
+    const folder = await prisma.folder.findFirst({
+      where: {
+        id: folderID,
+        expiresAt: { not: null },
+      },
+      include: {
+        files: true,
+        subFolders: true,
+      },
+    });
+
+    console.log("folder.expiresAt:", folder.expiresAt);
+    console.log("folder:", folder);
+    const expired = isExpired(folder.expiresAt);
+    console.log("expired:", expired);
+    if (expired) {
+      // Update folder's expiresAt column to null
+    }
+
+    const baseURL = "/share/";
+    // Need to prohibit specific buttons from rendering
+    res.render("folder", {
+      folder,
+      formAction: "",
+      baseURL,
+      public: true,
+    });
     // How to handle nested folders?
+    // Should the child folders only be accessible from the root>
     // Option 1
     //  Add a folderID parameter
     // Option 2
@@ -35,8 +67,11 @@ const shareController = {
       // console.log(
       //   Math.floor((expiresAt.getTime() - today.getTime()) / (1000 * 3600 * 24))
       // );
-      // traverseNestedFolders(folderID, updateFolderExpiresAt(expiresAt));
-      res.render("shareFolderOutput", { shareURL: "Hello world!", expiresAt });
+      traverseNestedFolders(folderID, updateFolderExpiresAt(expiresAt));
+      res.render("shareFolderOutput", {
+        shareURL: `/share/${folderID}`,
+        expiresAt,
+      });
       // res.status(200);
     }),
   ],