+++ utils/utils.js
@@ -2,6 +2,7 @@ const generateRandomIndex = require("./generateRandomIndex");
 const generateStoragePath = require("./generateStoragePath");
 const getFileExtension = require("./getFileExtension");
 const getFilesFromPath = require("./getFilesFromPath");
+const isExpired = require("./isExpired");
 const logger = require("./logger");
 
 module.exports = {
@@ -9,5 +10,6 @@ module.exports = {
   generateStoragePath,
   getFileExtension,
   getFilesFromPath,
+  isExpired,
   logger,
 };